rework LTRdigest
This commit splits the former GtLTRdigest stream into multiple individual
visitors, making them more modular and more in style with the rest of
the Gt annotation processing machinery. Moreover this code changes the protein
domain detection component of LTRdigest to call HMMER externally instead of
linking to it (licensing and redistributability reasons).